Splitted PointTasks into separate dart files
